+++ cardano-node/src/Cardano/Node/Tracing/API.hs
@@ -9,8 +9,8 @@ module Cardano.Node.Tracing.API
   ) where
 
 import           Cardano.Logging hiding (traceWith)
-import qualified Cardano.Logging.Types as Net
 import           Cardano.Logging.Prometheus.TCPServer (runPrometheusSimple)
+import qualified Cardano.Logging.Types as Net
 import           Cardano.Network.PeerSelection.PeerTrustable (PeerTrustable)
 import           Cardano.Node.Configuration.NodeAddress (PortNumber)
 import           Cardano.Node.Configuration.POM (NodeConfiguration (..))
@@ -51,6 +51,9 @@ import           Data.Time.Clock (getCurrentTime)
 import           Network.Mux.Trace (TraceLabelPeer (..))
 import           Network.Socket (HostName)
 import           System.Metrics as EKG
+
+import           Trace.Forward.Forwarding (initForwardingDelayed)
+import           Trace.Forward.Utils.TraceObject (writeToSink)
 
 
 initTraceDispatcher ::
@@ -137,7 +140,7 @@ initTraceDispatcher nc p networkMagic nodeKernel p2pMode noBlockForging = do
                 forwardingConf :: TraceOptionForwarder
                 forwardingConf = fromMaybe defaultForwarder (tcForwarder trConfig)
             initForwardingDelayed iomgr forwardingConf networkMagic (Just ekgStore) tracerSocketMode
-          pure (forwardTracer forwardSink, dataPointTracer dpStore, kickoffForwarder)
+          pure (forwardTracer (writeToSink forwardSink), dataPointTracer dpStore, kickoffForwarder)
         else
           -- Since 'Forwarder' backend isn't enabled, there is no forwarding.
           -- So we use nullTracers to ignore 'TraceObject's and 'DataPoint's.