+++ cardano-tracer/src/Cardano/Tracer/Handlers/ReForwarder.hs
@@ -15,7 +15,6 @@ module Cardano.Tracer.Handlers.ReForwarder
   ( initReForwarder
   ) where
 
-import           Cardano.Logging.Forwarding
 import           Cardano.Logging.Trace
 import           Cardano.Logging.Tracer.DataPoint
 import qualified Cardano.Logging.Types as Log
@@ -29,8 +28,9 @@ import           Control.Monad (when)
 import           Data.List (isPrefixOf)
 import qualified Data.Text as Text
 
-import           Trace.Forward.Utils.DataPoint
-import           Trace.Forward.Utils.TraceObject (ForwardSink, writeToSink)
+import           Trace.Forward.Forwarding
+import           Trace.Forward.Utils.ForwardSink (ForwardSink)
+import           Trace.Forward.Utils.TraceObject (writeToSink)
 
 -- | Initialize the reforwarding service if configured to be active.
 --   Returns